services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-fastapi-instrumentator==6.1.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6